Moved components to the quiz container
