+++ src/containers/quiz/points/Points.styles.ts
