+++ src/containers/quiz/quiz-info-section/QuizInfoSection.styles.ts
