+++ src/containers/quiz/quiz-info-section/QuizInfoSection.tsx
@@ -2,7 +2,7 @@ import Box from '@mui/material/Box'
 import Divider from '@mui/material/Divider'
 import Typography from '@mui/material/Typography'
 
-import styles from '~/components/quiz-info-section/QuizInfoSection.styles'
+import styles from '~/containers/quiz/quiz-info-section/QuizInfoSection.styles'
 
 type QuizInfoProps = {
   title: string
