+++ src/containers/quiz/timer/Timer.styles.ts
